requirements.txt
pylast>=5.3.0
tidalapi>=0.7.0
spotipy>=2.23.0
python-dotenv>=1.0.0
